dfdfdfdf

+++ pages/4 🏗️_apt_rebuild.py
@@ -328,7 +328,7 @@ if __name__ == "__main__":
         if response:
             st.write("선택한 아파트 위치:")
             selected_df = response["selected_rows"]
-
+            selected_df['공급면적'] = selected_df['공급면적'].astype(int)
             if selected_df:
                 px.set_mapbox_access_token(token)
                 fig = px.scatter_mapbox(selected_df, lat="위도", lon="경도", color="매물종류", size='공급면적', hover_name="단지명", hover_data=["특이사항", "한글거래가액", "시도명"],
